Add ORPL Routing Table tracking neighbours and calculating EqDC
Major component is the ORPLRoutingTable. This
uses the Algorithm for EDC from Opportunistic Routing in Low Duty-Cycle Wireless Sensor Networks - Ghadimi et al.
It implementation is similar to NextHopRoutingTable but it doesn't yet
implement the IRoutingTable Interface.
NextHopInterfaceData is added as an inteface tag so GlobalArp detects
and caches the wireless interface.

The p(i,j) is calculated differently since time and duty cycles cannot
be accurately tracked.
Instead Coincidental, Expected and no expected encounters signals are
sent from WakeUpMacLayer

The number of expected encounters from neighbors affects the probability
of reaching the neighbors. The main problem to address is what is the
denominator for the node encounters, since it cannot be the duty cycle.
This still needs work.

Statistics record the number and time of encounters and the change in
the EqDC.

Minor changes
* set default value for ExpectedCost to max by default
* Return EqDC when queryAcceptPacket to send indicator in Ack
* Move mac layer signal definition to within WakeUpMacLayer class.