Move nextHop EqDC to ORPLRoutingTable
Removing the ExpectedCostJump makes the implementation in line with the
proposed work. Instead the jump is represented by a new nextHop value
from calculateEqDC

Slight modification of MAC Acknowledgement so that duplicates are
reduced. This modification reduces ack collisions by limiting the
contention time and also stops transmission of Acks too late so the
transmitter will stop listening before the ack ends.